change to stand script
git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/tags/advLigoRTS-2.4@2730 6dcd42c9-f523-4c6d-aada-af552506706e

+++ etc/install/scripts/move_userapps.sh
@@ -4,16 +4,29 @@
 source /opt/cdscfg/rtsetup.sh
 #
 newHome=/opt/rtcds/userapps
-if [ -f ${newHome}]; then
+newPath=${newHome}/release
+if [ -d ${newHome} ]; then
     echo "Userapps already moved to /opt/rtcds/userapps"
+    cd ${newHome}    
 else
-    echo "Move userapps to /opt/rtcds/userapps"
-    mv /opt/rtcds/${site}/${ifo}/userapps ${newHome}
+    mkdir -p ${newHome}
+    oldHome=/opt/rtcds/${site}/${ifo}/userapps
+    if [ -d ${oldHome} ]; then
+       echo "Move userapps to /opt/rtcds/userapps"
+       mv /opt/rtcds/${site}/${ifo}/userapps/* ${newHome}
+       cd ${newHome}
+       rm release
+       ln -s trunk release
+    else
+       echo "No userapps found - check out a fresh one"
+       appsSvn="https://redoubt.ligo-wa.caltech.edu/svn/cds_user_apps"
+       cd ${newHome}
+       svn co ${appsSvn}
+       ln -s trunk release
+    fi
 fi
 echo "Update userapps from cds_user_apps Subversion"
 cd ${newHome}
-rm release
-ln -s trunk release
 svn update release
